requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
asyncpg
alembic
pydantic
pydantic-settings
pydantic[email]
httpx
python-jose[cryptography]
python-multipart
email-validator
passlib[bcrypt]
pytest
pytest-asyncio
aiosqlite
